requirements.txt
fastapi==0.111.0
mediapipe==0.10.8
numpy
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.3.0
Requests==2.32.2
torch==1.11.0
torchvision==0.12.0
webuiapi==0.9.9
